Time: 5 ms (35.85%), Space: 6 MB (53.14%) - LeetHub

+++ 1359-count-all-valid-pickup-and-delivery-options/1359-count-all-valid-pickup-and-delivery-options.cpp
@@ -1,14 +1,14 @@
 class Solution {
 public:
     int countOrders(int n) {
-        long long int dp;
+        int dp;
         dp = 1;
-        long long int mod = 1000000007;
+        int mod = 1000000007;
         for(int i =2; i<=n; i++)
         {
             long long int x = (i-1)*2+1;
             dp = (dp * ((x*(x+1))/2)) %mod;
         }
-        return (int)dp;
+        return dp;
     }
 };
